Refactor RegisterFunction type definition to auth.service.ts and update imports in related files
