+++ src/app/payment/services/auth-checkout.service.ts
@@ -1,5 +1,4 @@
-import { RegisterFunction } from 'app/auth/components/SignUp/useSignUp';
-import { logIn, signUp, SignUpParams } from 'app/auth/services/auth.service';
+import { logIn, RegisterFunction, signUp, SignUpParams } from 'app/auth/services/auth.service';
 import { AppDispatch } from 'app/store';
 import { AuthMethodTypes } from '../types';
